Se agrego atributo alerta a la entidad Documento

+++ src/java/co/com/gblock/ImplementacionServicios/VentaServicio.java
@@ -15,6 +15,6 @@ import javax.jws.WebService;
  */
 @Stateless
 @WebService(endpointInterface="co.com.gblock.InterfaceServicios.IDocumentoServicio")
-public class VentaServicio extends GenericaServicio<Documento, Integer> implements IDocumentoServicio{
+public class VentaServicio extends GenericaServicio<Documento, Long> implements IDocumentoServicio{
     
 }
